+++ include/fast_float/float_common.h
@@ -175,7 +175,7 @@ struct decimal {
   }
   // Generate san exponent matching to_truncated_mantissa()
   inline int32_t to_truncated_exponent() {
-    return decimal_point - max_digit_without_overflow;
+    return decimal_point - int32_t(max_digit_without_overflow);
   }
 };
